requirements.txt
Django==5.2.7
djangorestframework==3.15.2
requests==2.31.0
